% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIpInfo.R
\name{get_ip_info}
\alias{get_ip_info}
\title{Checks the provided IP address (both v4 & v6 formats) and returns all available information.}
\usage{
get_ip_info(ip = NULL, language = NULL)
}
\arguments{
\item{ip}{The IP address you want to query}

\item{language}{An ISO Alpha 2 Language Code for localizing the IP data}
}
\value{
Returns all available information about an IP address.
}
\description{
Checks the provided IP address (both v4 & v6 formats) and returns all available information.
}
